Scale animation's texture during load
During the loading of an animation, SCALE values is read from the file and the animation's texture is scaled
accordingly. Problem may arise when the texture file is not loaded correctly and Texture object doesn't exist.
Also in Animation::reset() method we match the dest_rect dimension with those of src_rect. This should be done
also during load.