Add Polygon class and loading polygons funnctionality for Room
Polygon is simply a vector of points (Vector2D object, just x and y). It has simple methods . One for adding a vertex, which is used during loading.
Also methods for scaling and moving, which will be used for drawing polygons. Drawing of polygons will be implemented only for testing purposes.

Another, very important method, is point_in_polygon. This method returns true if a given point is inside the polygon.

Added "walkarea" field into rooms json, which is loaded into the Room object.