Make walk_area non-pointer field
When to use pointers? That is the question. When a object is created on the heap, all its fields are probably also on the heap. So now the walk_area is not a pointer,but a regular field.

The renderable PolygonObject was adapted to this change. It now has a reference to the polygon it should visualize and must be given a scale and dx, dy values to scale and move final image of the polygon. This scaling is done during rendering.