Use smart unique pointer in the Room class.
Room's texture, screen_walk_area, and click_map_texture are now unique_ptrs. Renderer's renderable objects queue
still hold normal pointers. Renderer just uses them, Room controls their lifespan.

The loading of the rooms changed a bit because of the smart pointers. Unique_ptr cannot be copied, therefore it
cannot be passed as an argument by value. It can only be passed as reference or moved. I don't know how to use
move in this situation. So a created an empty contructor for Room. This will automatically set unique_ptrs to
nullptr (walk area and click map renderables therefore don't have to be set to nullptr manually). After this
initialization a load method is called. This method populates all the Room's fields.

The RenderManager::register_object method takes raw pointers. Why? Because unique_ptr cannot be copied. So we
pass the smart pointer's underlying raw pointer using the .get() method.

The RenderManager::scale_full_h method expected a reference to a Texture object. Don't know how to get that from
a smart pointer. So as a workaround overloaded ersion of this method, which takes in a const reference to a
unique_ptr was created. Is there a better way?

Animations do not work properly now. Animation's texture should also be changet to unique_ptr.