Use smart pointer in Animation class
The texture in Animation is now also a unique_ptr. The loading of animations was changed. Just like while loading rooms, first a
vector of empty animations is created and then populated. Animation therefore has a new move constructor.

The timing of the animations is brokem. Animations seem to dissapear on some frames, and are generally too fast and clunky.