Use size_type for animation vector iteration
size_type seems to be more generic and portable, it depends on the container, and is infact part of the container namespace.

I used "alias", not sure what the proper term is, to define size_type for my vector of animations. It's not that helpful in this case, but
I wanted to try it.